Block Comments: Extract current data querying logic into a hook
